refactor: constant 파일 정리 몇몇 파일 common에 통합
