--- src/constants/toast.constant.ts
@@ -1,1 +1,0 @@
-export const TOAST_ID = 'toast';
