+++ src/hooks/useToast.ts
@@ -3,7 +3,7 @@ import { useSetRecoilState } from 'recoil';
 import { toastState } from '@utils/atoms/common.atom';
 import { ToastType } from '@utils/types/common.type';
 
-import { TOAST_ID } from '@constants/toast.constant';
+import { TOAST_ID } from '@constants/common.constant';
 
 const useToast = () => {
   const setToastState = useSetRecoilState(toastState);
