+++ src/utils/atoms/common.atom.ts
@@ -3,7 +3,7 @@ import { atom, selector } from 'recoil';
 import { ModalType, toastStateType } from '@utils/types/common.type';
 
 import { ATOM_KEY } from '@constants/atom.constant';
-import { TOAST_ID } from '@constants/toast.constant';
+import { TOAST_ID } from '@constants/common.constant';
 
 export const modalState = atom<ModalType>({
   key: ATOM_KEY.MODAL,
